split up groups, tasks, and members files into more readable seperate components and changed folder structure
