+++ src/App.js
@@ -10,12 +10,12 @@ import theme from "./styles/theme/index"
 import Dashboard from "./components/views/Dashboard"
 // import Login from './components/views/Login'
 import CreateFamily from "./components/views/CreateFamily"
-import Groups from "./components/views/Groups"
-import Group from "./components/views/Group"
+import Groups from "./components/Group/Groups"
+import Group from "./components/Group/Group"
 import CreateAccount from "./components/views/CreateAccount"
-import Members from "./components/views/Members"
-import Member from "./components/views/Member"
-import Task from "./components/views/Task"
+import Members from "./components/Members/Members"
+import Member from "./components/Members/Member"
+import Task from "./components/task/Task"
 import Settings from "./components/views/Settings"
 import styled, { ThemeProvider, css } from "styled-components"
 import "./styles/theme/global"
