+++ src/components/views/Dashboard.js
@@ -1,7 +1,7 @@
-import React from 'react'
-import Action from '../cards/Action'
-import Notification from '../shared/Notification'
-import ContentWrapper from '../../styles/ContentWrapper'
+import React from "react"
+import Action from "../shared/ActionCard"
+import Notification from "../shared/Notification"
+import ContentWrapper from "../../styles/ContentWrapper"
 
 class Dashboard extends React.Component {
   render() {
